Add files via upload
This repository contains my solutions to various hackathons organized by Machinehack. Each folder within the repository corresponds to a specific hackathon, featuring the problem statement, data, and my approach to solving the problem. The solutions showcase a range of techniques and methodologies used in data science and machine learning to tackle